CreatedUserDetailUpdateMethod

+++ Shop.Business/Services/AdminServices.cs
@@ -14,48 +14,65 @@ public class AdminServices : IAdminServices
     }
     public void AddDiscountToProduct(int productId, int discountId)
     {
-        var pro = context.Products.Find(productId);
-        if (pro is null) throw new DoesNotExistException($"Product with Id :{productId} doesn't exist");
-        var dis = context.Discounts.Find(discountId);
-        if (dis is null) throw new DoesNotExistException($"Discount with Id :{discountId} doesn't exist");
-        if (pro.DiscountId != null) throw new AlreadyExistException($"This product already has a discount");
-        if (pro.DiscountId == discountId) throw new AlreadyExistException($"This product already has this discount");
-        if (dis.IsActive == false) throw new DoesNotExistException($"Discount with Id :{discountId} disabled");
-        pro.DiscountId = discountId;
-        pro.Price = pro.Price - (pro.Price * dis.Percentage / 100);
-        context.SaveChanges();
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
+        {
+            var pro = context.Products.Find(productId);
+            if (pro is null) throw new DoesNotExistException($"Product with Id :{productId} doesn't exist");
+            var dis = context.Discounts.Find(discountId);
+            if (dis is null) throw new DoesNotExistException($"Discount with Id :{discountId} doesn't exist");
+            if (pro.DiscountId != null) throw new AlreadyExistException($"This product already has a discount");
+            if (pro.DiscountId == discountId) throw new AlreadyExistException($"This product already has this discount");
+            if (dis.IsActive == false) throw new DoesNotExistException($"Discount with Id :{discountId} disabled");
+            pro.DiscountId = discountId;
+            pro.Price = pro.Price - (pro.Price * dis.Percentage / 100);
+            context.SaveChanges();
+
+        }
 
     }
 
     public void CreateDiscount(string name, int percentage)
     {
-        var dis = context.Discounts.FirstOrDefault(d => d.Name == name);
-        if (dis is not null) throw new AlreadyExistException($"Discount with name {name} already is exist");
-        Discount discount = new()
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            Name = name,
-            Percentage = percentage
+            var dis = context.Discounts.FirstOrDefault(d => d.Name == name);
+            if (dis is not null) throw new AlreadyExistException($"Discount with name {name} already is exist");
+            Discount discount = new()
+            {
+                Name = name,
+                Percentage = percentage
 
-        };
-        discount.IsActive = true;
-        context.Discounts.Add(discount);
-        context.SaveChanges();
+            };
+            discount.IsActive = true;
+            context.Discounts.Add(discount);
+            context.SaveChanges();
+
+        }
     }
 
     public void CreateProduct(string name, decimal price, int discountId, int proCount)
     {
-
-        var pro = context.Products.FirstOrDefault(p => p.Name.ToLower() == name.ToLower());
-        if (pro is not null) throw new AlreadyExistException($"Product with name :{name} already is exist");
-        Product product = new()
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            Name = name,
-            Price = price,
-            DiscountId = discountId,
-            ProductCount = proCount
-        };
-        context.Products.Add(product);
-        context.SaveChanges();
+            var pro = context.Products.FirstOrDefault(p => p.Name.ToLower() == name.ToLower());
+            if (pro is not null) throw new AlreadyExistException($"Product with name :{name} already is exist");
+            Product product = new()
+            {
+                Name = name,
+                Price = price,
+                DiscountId = discountId,
+                ProductCount = proCount
+            };
+            context.Products.Add(product);
+            context.SaveChanges();
+
+        }
 
 
 
@@ -63,40 +80,58 @@ public class AdminServices : IAdminServices
 
     public void DeleteProduct(int productId)
     {
-        var pro = context.Products.Find(productId);
-        if (pro is null) throw new DoesNotExistException($"Product with Id :{productId} doesn't exist");
-        if (pro.ProductCount > 0) throw new StillHasProCountInStockException($"Product with Id :{productId} still exist in product stock");
-        context.Products.Remove(pro);
-        context.SaveChanges();
-           
-        
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
+        {
+            var pro = context.Products.Find(productId);
+            if (pro is null) throw new DoesNotExistException($"Product with Id :{productId} doesn't exist");
+            if (pro.ProductCount > 0) throw new StillHasProCountInStockException($"Product with Id :{productId} still exist in product stock");
+            context.Products.Remove(pro);
+            context.SaveChanges();
+
+        }
+
+
     }
 
     public void DisableDiscount(int discountId)
     {
-        var dis = context.Discounts.Find(discountId);
-        if(dis is null) throw new DoesNotExistException($"Discount with Id :{discountId} doesn't exist");
-        if (dis.IsActive == false) throw new AlreadyDoesNotActiveException($"Discount with Id :{discountId} already deactivated");
-        dis.IsActive = false;
-        foreach (var product in context.Products)
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            if(product.DiscountId == discountId)
+            var dis = context.Discounts.Find(discountId);
+            if (dis is null) throw new DoesNotExistException($"Discount with Id :{discountId} doesn't exist");
+            if (dis.IsActive == false) throw new AlreadyDoesNotActiveException($"Discount with Id :{discountId} already deactivated");
+            dis.IsActive = false;
+            foreach (var product in context.Products)
             {
-                product.DiscountId = null;
-                product.Price = product.Price + (product.Price * dis.Percentage / 100);
+                if (product.DiscountId == discountId)
+                {
+                    product.DiscountId = null;
+                    product.Price = product.Price + (product.Price * dis.Percentage / 100);
 
+                }
             }
+            context.SaveChanges();
+
         }
-        context.SaveChanges();
     }
 
-    
+
 
     public void ShowAllBaskets()
     {
-        foreach(var basket in context.Baskets)
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            Console.WriteLine($"Basket Id:{basket.Id}; Basket's userId{basket.UserId}; Basket's product count:{basket.ProductCount}");
+            foreach (var basket in context.Baskets)
+            {
+                Console.WriteLine($"Basket Id:{basket.Id}; Basket's userId{basket.UserId}; Basket's product count:{basket.ProductCount}");
+
+            }
 
         }
     }
@@ -108,7 +143,7 @@ public class AdminServices : IAdminServices
             var dis = context.Discounts.Find(product.DiscountId);
             if (dis is null)
             {
-            Console.WriteLine($"Product Id :{product.Id} ; product name:{product.Name} ; product price: {product.Price}");
+                Console.WriteLine($"Product Id :{product.Id} ; product name:{product.Name} ; product price: {product.Price}");
 
             }
             else
@@ -124,17 +159,67 @@ public class AdminServices : IAdminServices
 
     public void ShowAllUsers()
     {
-        foreach( var user in context.Users)
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            Console.WriteLine($"User Id:{user.Id}; user name{user.Username} ; ");
-        };
+            foreach (var user in context.Users)
+            {
+                Console.WriteLine($"User Id:{user.Id}; user name{user.Username} ; ");
+            };
+
+        }
     }
 
     public void ShowAllWallets()
     {
-        foreach(var wallet in context.Wallets)
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new NotLoggedInException("Login as Admin user to choose this operation ");
+        else
         {
-            Console.WriteLine($"Wallet Id:{wallet.Id} wallet's userId:{wallet.UserId}; CardNumber: {wallet.CardNumber}; cardBalance: {wallet.CardBalance}");
+
+            foreach (var wallet in context.Wallets)
+            {
+                Console.WriteLine($"Wallet Id:{wallet.Id} wallet's userId:{wallet.UserId}; CardNumber: {wallet.CardNumber}; cardBalance: {wallet.CardBalance}");
+            }
         }
     }
+
+    public void AdminUserLogin(string userName, string email, string password)
+    {
+        var u1 = context.Users.FirstOrDefault(u => u.Username == userName);
+        if (u1 is null)
+        {
+            var u2 = context.Users.FirstOrDefault(u1 => u1.Email == email);
+            if (u2 is null) throw new DoesNotExistException($"Admin user with name:{userName} and email:{email} doesn't exist");
+            if (u2.Password == password && u2.IsAdminUser == true)
+            {
+                u2.ARegistr = true;
+            }
+            context.SaveChanges();
+
+        }
+        else
+        {
+            if (u1.Password == password && u1.IsAdminUser == true)
+            {
+                u1.ARegistr = true;
+            }
+            context.SaveChanges();
+        }
+
+
+    }
+
+    public void AdminUserLogout()
+    {
+        var us = context.Users.FirstOrDefault(u => u.ARegistr == true);
+        if (us is null) throw new DoesNotExistException("You need to log in as adminUser to log out");
+        us.ARegistr = false;
+        context.SaveChanges();
+
+
+    }
+
+
 }
